Fix: Code check
I check the password generator and now the method used an int as the
lenght of the password

+++ lib/pages/home_page.dart
@@ -215,7 +215,7 @@ class _HomePageState extends State<HomePage> {
                           onChanged: (double value) {
                             setState(() {
                               passwordStore
-                                ..changeLenght(value)
+                                ..changeLenght(value.toInt())
                                 ..generateNew();
                             });
                           })),
